FIX rename ramp_custom->ramp_custom_lwfa
